Refactored to use feature based directories
