+++ frontend/src/games/components/GameBanner.js
