+++ frontend/src/games/components/GameCard.js
