+++ frontend/src/games/components/SkeletonCard.js
