+++ frontend/src/performances/components/AdditionalInfo.js
