+++ frontend/src/setup/api/hardware.js
