+++ frontend/src/setup/components/RamSelection.js
