+++ frontend/src/setup/hooks/useHardwareSearch.js
